move stuff to godotsteam loader
